some logic to compress-and-archive to check for files

+++ land-spinup/inline-dates/compress-and-archive.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-CASESTR="ERA5-ELMinic-conus256_20020101_000_01"
+CASESTR="ERA5-ELMinic-ne128pg2_v2_20020101_000_01"
 DIRECTORY="/pscratch/sd/c/czarzyck/e3sm_scratch/pm-cpu/${CASESTR}/run/"
 IGNORELAST=1
-FINALDIR="/global/cfs/cdirs/m2637/E3SM_SCREAM_files/finidat/conustight256x8/ERA5/"
+FINALDIR="/global/cfs/cdirs/m2637/E3SM_SCREAM_files/finidat/ne128pg2_v2/ERA5/"
 
 #####
 
@@ -25,20 +25,29 @@ if (( count <= IGNORELAST )); then
   exit 0
 fi
 
-# Loop through all but last 2
+# Loop through all but last $IGNORELAST
 for ((i=0; i<count-$IGNORELAST; i++)); do
   f=${files[$i]}
   timestring="${f##*.elm.r.}"  # gets 2002-09-23-43200.nc
   timestring="${timestring%.nc}"  # strips off .nc
   echo $timestring
-  compress_file "${CASESTR}.elm.r.${timestring}.nc" zstd
-  compress_file "${CASESTR}.mosart.r.${timestring}.nc" zstd
-  rm -fv "${CASESTR}.mosart.rh0.${timestring}.nc"
-  rm -fv "${CASESTR}.elm.rh0.${timestring}.nc"
+  # Check if file exists first before compressing or removing
+  [[ -f "${CASESTR}.elm.r.${timestring}.nc" ]] && compress_file "${CASESTR}.elm.r.${timestring}.nc" zstd
+  [[ -f "${CASESTR}.mosart.r.${timestring}.nc" ]] && compress_file "${CASESTR}.mosart.r.${timestring}.nc" zstd
+  [[ -f "${CASESTR}.mosart.rh0.${timestring}.nc" ]] && rm -fv "${CASESTR}.mosart.rh0.${timestring}.nc"
+  [[ -f "${CASESTR}.elm.rh0.${timestring}.nc" ]] && rm -fv "${CASESTR}.elm.rh0.${timestring}.nc"
 done
 
-mv -v ${CASESTR}.elm.r.*.zst $FINALDIR
-mv -v ${CASESTR}.mosart.r.*.zst $FINALDIR
+# Create archival dir if it doesn't exist
+mkdir -p $FINALDIR
+
+# compgen -G checks to see if any files exist first, if they do move.
+if compgen -G "${CASESTR}.elm.r.*.zst" > /dev/null; then
+  mv -v ${CASESTR}.elm.r.*.zst "$FINALDIR"
+fi
+if compgen -G "${CASESTR}.mosart.r.*.zst" > /dev/null; then
+  mv -v ${CASESTR}.mosart.r.*.zst "$FINALDIR"
+fi
 
 # Calculate and display completion time and duration
 END_TIME=$(date +%s)
